rename packages in Android project
