+++ Android/skeleton-extension-android/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.sample.acp.extension"
+        applicationId "com.sample.company.tester"
         minSdkVersion 24
         targetSdkVersion 28
         versionCode 1
