+++ Android/skeleton-extension-android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.sample.acp.extension">
+    package="com.sample.company.tester">
 
     <application
         android:allowBackup="true"
