+++ Android/skeleton-extension-android/app/src/test/java/com/sample/company/tester/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.sample.acp.extension;
+package com.sample.company.tester;
 
 import org.junit.Test;
 
@@ -12,6 +12,6 @@ import static org.junit.Assert.*;
 public class ExampleUnitTest {
     @Test
     public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
+        Assert.assertEquals(4, 2 + 2);
     }
 }
