+++ Android/skeleton-extension-android/skeletonextension/build.gradle
@@ -3,8 +3,6 @@ apply plugin: 'com.android.library'
 android {
     compileSdkVersion 28
 
-
-
     defaultConfig {
         minSdkVersion 24
         targetSdkVersion 28
