+++ Android/skeleton-extension-android/skeletonextension/src/androidTest/java/com/sample/company/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.sample.acp.extension;
+package com.sample.company;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.sample.acp.extension", appContext.getPackageName());
+        assertEquals("com.sample.extension.test", appContext.getPackageName());
     }
 }
