+++ Android/skeleton-extension-android/skeletonextension/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.sample.acp.extension" />
+    package="com.sample.company.extension" />
